+++ src/data/match.ts
@@ -1,11 +1,11 @@
 export const MATCH_PAGE_DATA = [
-    { id: 1, pathname: 'match/type', width: '1206px', height: '227px', top: '36px', left: '41px', title: '바로 매칭하기' },
+    { id: 1, pathname: 'match', width: '1206px', height: '227px', top: '36px', left: '41px', title: '바로 매칭하기' },
     { id: 2, pathname: 'board', width: '1206px', height: '227px', top: '36px', left: '41px', title: '매칭 게시판에서 찾기' },
 ];
 
 export const MATCH_TYPE_PAGE_DATA = [
-    { id: 1, type:'fun',pathname: '/game-mode', width: '600px', height: '380px', top: '50%', left: '50%', title: '즐겜' },
-    { id: 2, type:'hard',pathname: '/game-mode', width: '600px', height: '380px', top: '50%', left: '50%', title: '빡겜' },
+    { id: 1, type:'fun',pathname: 'match/game-mode', width: '600px', height: '380px', top: '50%', left: '50%', title: '즐겜' },
+    { id: 2, type:'hard',pathname: 'match/game-mode', width: '600px', height: '380px', top: '50%', left: '50%', title: '빡겜' },
 ];
 
 export const GAME_MODE_PAGE_DATA = [
